requirements.txt
selenium
webdriver-manager
boto3
python-dotenv
pandas
scikit-learn
joblib
